Quasi-Newton: evaluate at predicted minimum rather than at coordinate-axis approximation
Previously there was unnecessary backtracking during line search, because we weren't getting the full benefit of the Newton solution. The issue was that forcing the evaluation point to lie on a line with xnew[i] == xold[i] for all but 1 i typically gave up too much of the benefit of the true target to lead to an improvement in function value.

The downside here is that now there are function evaluations that never get recorded in the tree structure.