Provide for immediate termination based on fvalue
Formerly, the `fvalue` termination criterion was tested after the end
of a sweep, even if that criterion was satisfied before processing the
entire sweep queue. When the dimensionality `N` is high, this can lead
to a considerable delay, as the number of queued items is likely to be
proportional to `N` and the cost of quasi-Newton iteration goes like
`N^3`. Since even a single quasi-Newton step can dramatically decrease
the objective value (possibly below `fvalue`), it's better to test
the `fvalue` criterion after each point is processed.