Move quasinewton after sweep, add convexity test, more stability
This cleans up code a lot by separating the DIRECT portion of the
algorithm from the quasi-Newton step. This gives us more analysis
opportunities without having to pass even more parameters to
autosplit!. Any box that gets split gets queued on a list of potential
quasi-Newton "seeds."

I exploit this by doing a convexity analysis on candidate seeds. The
idea is to look at boxes that lie between the seeds and determine
whether there is a peak in the function value on the line between
them. This drastically cuts down on the number of quasi-Newton
steps attempted.

This also marks boxes that attempted a QN step but failed to improve
by enough to exceed the convergence tolerance. This reduces
repeated work.

Finally, when building the quadratic model it now also checks the
condition number of Gaussian elimination, bringing extra points
until the condition number seems reasonable. I discovered this
to be necessary when I fixed a bug in the high-dimensional test,
which had inadvertently constructed a diagonal Hessian and
consequently didn't need the quasi-Newton step for
convergence.