Move quasinewton after sweep, force completion of quadratic model
This cleans up code a lot by separating the DIRECT portion of the
algorithm from the quasi-Newton step. This gives us more analysis
opportunities without having to pass even more parameters to
autosplit!. Any box that gets split gets queued on a list of potential
quasi-Newton "seeds."

This also implements the changes needed to finalize the quadratic model.
It selects points for evaluation that fill in missing information about
the regression coefficients. The point selection is based largely
on Gaussian elimination to ensure independence. Unfortunately due
to the box-splitting it needs to evaluate 2 points for every extra
piece of new information about the regression coefficients.

This also marks boxes that attempted a QN step but failed to improve
by enough to exceed the convergence tolerance. It also excludes later
points that fit the local quadratic model. Together these reduce
repeated work. It also fixes a number of other issues with QN backtracking,
and adds scaling to improve numeric stability.